[BACKPORT 2.11.1][PLAT-2673] Validating custom keypair with AWS
Summary:
Allows custom keypair to already exist on AWS by validating the fingerprint.
Validates using either MD5 or SHA1 fingerprint checksum.
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-key-pairs.html#verify-key-pair-fingerprints
Does not interrupt current flow where custom keypair can be created with private key content.
Also adds an override button that will not perform any validation (to avoid AWS permission issues) nor will it try to delete the key from AWS.

Original diff: https://phabricator.dev.yugabyte.com/D14716

Test Plan:
Jenkins: rebase: 2.11.1

Create provider with custom keypair that already exists in AWS.
Verify universe can be created with provider.
Repeat for keypair that is imported to AWS (yb-dev-aws-2), created in AWS (mchidambaram-keypair),
and created via ssh-keygen and then imported (mchidambaram-openssh).
Verify new provider can be created with custom content (mchidambaram-dev-aws-2) and universe can be
created.
Verify if incorrect private key content is provided with keypair name that exists in AWS error is
raised.

Reviewers: streddy, nbhatia, alan, nsingh

Reviewed By: nsingh

Subscribers: yugaware, jenkins-bot

Differential Revision: https://phabricator.dev.yugabyte.com/D15117

+++ managed/devops/opscli/ybops/cloud/common/method.py
@@ -1154,6 +1154,7 @@ class AbstractAccessMethod(AbstractMethod):
         self.parser.add_argument("--key_file_path", required=True, help="Key file path")
         self.parser.add_argument("--public_key_file", required=False, help="Public key filename")
         self.parser.add_argument("--private_key_file", required=False, help="Private key filename")
+        self.parser.add_argument("--delete_remote", action="store_true")
 
     def validate_key_files(self, args):
         public_key_file = args.public_key_file
@@ -1199,7 +1200,8 @@ class AccessDeleteKeyMethod(AbstractAccessMethod):
 
     def callback(self, args):
         try:
-            self._delete_key_pair(args)
+            if (args.delete_remote):
+                self._delete_key_pair(args)
             self._cleanup_dir(args.key_file_path)
             print(json.dumps({"success": "Keypair {} deleted.".format(args.key_pair_name)}))
         except Exception as e:
